# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaterOS SaaS API Documentation",
        "description": "Dokumentasi RESTful API untuk platform SaaS Management Catering End-to-End CaterOS.",
        "contact": {
            "name": "CaterOS Engineering Team",
            "email": "dev@cateros.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/tenant/menu-packages": {
            "get": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Daftar Paket Menu & Bundling",
                "description": "Get all menu packages with bundled items and HPP.",
                "operationId": "897053ac1bdb80956a951744ee1aa677",
                "parameters": [
                    {
                        "name": "package_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar paket menu berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Tambah Paket Menu Baru (Bundling)",
                "description": "Store new menu package with bundled items.",
                "operationId": "d9dd510c5e008e55cad8dd956e360af5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "package_type",
                                    "selling_price",
                                    "items"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paket Nasi Kotak Komplit A"
                                    },
                                    "package_type": {
                                        "type": "string",
                                        "example": "nasi_kotak"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "PKG-NK-01"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nasi Putih + Ayam Bakar + Sambal Goreng Kentang + Sayur + Kerupuk"
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "example": 35000
                                    },
                                    "min_order_quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_item_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "Paha / Dada"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paket menu berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/menu-packages/{id}": {
            "get": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Detail Paket Menu Bundling",
                "description": "Show menu package detail.",
                "operationId": "964fc59a2804c7d31d1ac69661ee0594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail paket berhasil diambil"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Update Paket Menu & Item Bundling",
                "description": "Update menu package.",
                "operationId": "e5ca1ecd27b86a3dd40856989180f6a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Paket Nasi Kotak Super Komplit"
                                    },
                                    "selling_price": {
                                        "type": "number",
                                        "example": 38000
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_item_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paket menu berhasil diperbarui"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Hapus Paket Menu",
                "description": "Delete menu package.",
                "operationId": "a346637df70344ab60065b81e7202bd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paket berhasil dihapus"
                    },
                    "404": {
                        "description": "Paket tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
        "/tenant/menu-packages/{id}/image": {
            "post": {
                "tags": [
                    "Master Paket Menu (Bundling)"
                ],
                "summary": "Upload Foto Paket Menu",
                "description": "Upload photo for menu package.",
                "operationId": "96f62fa869445e89a3dd4f213fd24a00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foto paket berhasil diunggah"
                    },
                    "422": {
                        "description": "Validasi gambar gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "TenantHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer Sanctum (contoh: 1|abcdef...)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "TenantHeader": {
                "type": "apiKey",
                "description": "Header ID Tenant yang sedang aktif (contoh: 1)",
                "name": "X-Tenant-ID",
                "in": "header"
            }
        }
    },
...
}